requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.13.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py37h271585c_0
albumentations=1.0.3=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py37h271585c_1001
bzip2=1.0.8=h0d85af4_4
c-ares=1.17.1=h0d85af4_1
ca-certificates=2021.5.30=h033912b_0
cachetools=4.2.2=pyhd8ed1ab_0
cairo=1.16.0=h0ab9d94_1001
certifi=2021.5.30=py37hf985489_0
cffi=1.14.6=py37hda351d9_0
chardet=4.0.0=py37hf985489_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=py37hf985489_0
coverage=5.5=py37h271585c_0
cryptography=3.4.7=py37hce4a858_0
cycler=0.10.0=pypi_0
dataclasses=0.6=pypi_0
decorator=4.4.2=pypi_0
efficientnet-pytorch=0.6.3=pypi_0
ffmpeg=4.0=h01ea3c9_0
flask=2.0.1=pypi_0
fontconfig=2.13.1=h10f422b_1005
freetype=2.10.4=h4cff582_1
future=0.18.2=pypi_0
gettext=0.21.0=h7535e17_0
glib=2.68.2=hdf23fa2_0
google-auth=1.33.1=pyh6c4a22f_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.13=h12caacf_1001
grpcio=1.35.0=py37h97de6d8_0
harfbuzz=1.8.8=hb8d4a28_0
hdf5=1.10.2=hc401514_3
icu=58.2=h0a44026_1000
idna=3.1=pyhd3deb0d_0
imageio=2.9.0=pypi_0
importlib-metadata=4.6.1=py37hf985489_0
imutils=0.5.4=pypi_0
intel-openmp=2021.2.0=hecd8cb5_564
itsdangerous=2.0.1=pypi_0
jasper=2.0.14=hee3b9ff_2
jinja2=3.0.1=pypi_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=hbcb3906_0
kiwisolver=1.3.1=pypi_0
lcms2=2.12=hf1fd2bf_0
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=0
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=9_mkl
libopencv=3.4.2=h7c891bd_1
libopus=1.3.1=hc929b4f_1
libpng=1.6.37=h7cec526_2
libprotobuf=3.14.0=h2842e9f_0
libtiff=4.2.0=h87d7836_0
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=11.1.0=hda6cdc1_1
lz4-c=1.9.3=h23ab428_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.2=pypi_0
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.4.0=py37h271585c_0
mkl_fft=1.3.0=py37h4a7008c_2
mkl_random=1.2.1=py37hb2f4e1b_2
multidict=5.1.0=py37h271585c_1
munch=2.5.0=pypi_0
natsort=7.1.1=pypi_0
ncurses=6.2=h0a44026_1
networkx=2.5.1=pypi_0
ninja=1.10.2=hf7b0b51_1
numpy=1.20.2=py37h4b4dc7a_0
numpy-base=1.20.2=py37he0bd621_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
opencv-python-headless=4.5.2.54=pypi_0
openssl=1.1.1k=h0d85af4_0
pandas=1.2.5=py37h23ab428_0
pcre=8.45=h23ab428_0
pillow=8.3.0=pypi_0
pip=21.1.3=py37hecd8cb5_0
pixman=0.38.0=h01d97ff_1003
pretrainedmodels=0.7.4=pypi_0
protobuf=3.14.0=py37h23ab428_1
py-opencv=3.4.2=py37h7c891bd_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=2.1.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=py37hf985489_3
python=3.7.10=h88f2d9e_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.7.0=py3.7_0
pytz=2021.1=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1=h9ed2024_0
requests=2.26.0=pyhd8ed1ab_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
scikit-image=0.18.2=pypi_0
scikit-learn=0.22.1=py37h3dc85bc_1
scipy=1.7.0=pypi_0
segmentation-models-pytorch=0.2.0=pypi_0
setuptools=52.0.0=py37hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=hce871da_0
tensorboard=2.5.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py37hf4a05c0_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tifffile=2021.7.2=pypi_0
timm=0.4.12=pypi_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.7.0=py37
torchvision=0.8.1=py37_cpu
tqdm=4.61.2=pypi_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yarl=1.6.3=py37h271585c_2
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
